Added code to handle the URL scheme callback mechanism for the Facebook login credentials
Apparently, the Facebook login mechanism is to launch either the Facebook app,
if installed on the device, or Safari to do a web login, and once the login is
complete, our application will receive the login credentials by having
registered to handle a URL scheme which gets opened.  This version of the code
doesn't seem to actually work, however, so more effort is going to be needed
here.